refactor: 런닝 api 내부 디렉토리 추가 #168
