+++ src/main/java/com/runky/running/domain/RunningService.java
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.runky.global.error.GlobalException;
-import com.runky.running.api.WsDestinations;
+import com.runky.running.api.socket.WsDestinations;
 import com.runky.running.error.RunningErrorCode;
 
 import lombok.RequiredArgsConstructor;
