+++ src/test/java/com/runky/running/api/RunningApiE2ETest.java
@@ -19,6 +19,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 import com.runky.global.response.ApiResponse;
+import com.runky.running.api.http.RunningRequest;
+import com.runky.running.api.http.RunningResponse;
 import com.runky.running.domain.Running;
 import com.runky.running.domain.RunningRepository;
 import com.runky.running.domain.RunningTrack;
